% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_test.R
\name{ks.test.fitdist}
\alias{ks.test.fitdist}
\title{\link[stats]{ks.test} on \link[fitdistrplus]{fitdist}}
\usage{
\method{ks.test}{fitdist}(x, y = paste0("p", x$distname), data = x$data, ...)
}
\arguments{
\item{x}{a \link[fitdistrplus]{fitdist} object}

\item{y}{\link[base]{character} scalar, the cumulative distribution function}

\item{data}{\link[base]{numeric} \link[base]{vector}}

\item{...}{parameters of the cumulative distribution. Defaults are
the estimate from the \link[fitdistrplus]{fitdist} object}
}
\description{
\link[stats]{ks.test} on \link[fitdistrplus]{fitdist}
}
\examples{
library(fitdistrplus)
data(groundbeef, package = 'fitdistrplus')
groundbeef$serving |> 
 fitdist(distr = 'gamma') |>
 ks.test() |> 
 suppressWarnings()
}
